refactor: rename helpers folder and CRUDOperations file
